% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqlite.R
\name{unqlite_close}
\alias{unqlite_close}
\title{Close a database connection}
\usage{
unqlite_close(db)
}
\arguments{
\item{db}{UnqliteConnection object to be closed}
}
\description{
Close a database connection
}
