% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqlite.R
\name{unqlite_kv_delete}
\alias{unqlite_kv_delete}
\title{Delete an entry in the database}
\usage{
unqlite_kv_delete(db, key)
}
\arguments{
\item{db}{UnqliteConnection object}

\item{key}{Any R object, if the key does not exist, nothing is done}
}
\description{
Delete an entry in the database
}
