% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqlite.R
\name{unqlite_kv_fetch}
\alias{unqlite_kv_fetch}
\title{Fetch an entry from a database}
\usage{
unqlite_kv_fetch(db, key)
}
\arguments{
\item{db}{UnqliteConnection object}

\item{key}{Any R object}
}
\description{
Fetch an entry from a database
}
