% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqlite.R
\name{unqlite_kv_store}
\alias{unqlite_kv_store}
\title{Store a key/value pair}
\usage{
unqlite_kv_store(db, key, val)
}
\arguments{
\item{db}{UnqliteConnection object}

\item{key}{Any R object}

\item{value}{Any R object}
}
\description{
Store a key/value pair
}
