% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runqlite.R
\name{unqlite_open}
\alias{unqlite_open}
\title{Open/create a new UnQLite database}
\usage{
unqlite_open(filename)
}
\arguments{
\item{filename}{UnQLite database filename}
}
\value{
UnqliteConnection object
}
\description{
An UnQLite database stores data in a local file. This function opens such a
database (or creates it, if no file is found). This file will persist across
sessions.
}
