% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{UnqliteConnection-class}
\alias{UnqliteConnection-class}
\alias{UnqliteConnection}
\title{UnQLite database object}
\description{
This class holds an external pointer to an UnQLite database and the current
cursor position. You should never need to do anything with its internal
values.
}
\section{Slots}{

\describe{
\item{\code{dbptr}}{External pointer to the UnQLite database}

\item{\code{cursor}}{Cursor position}
}}

